+++ libtock/ipc.c
@@ -1,7 +1,7 @@
 #include "ipc.h"
 #include "tock.h"
 
-int ipc_discover(const char* pkg_name, int* svc_id) {
+int ipc_discover(const char* pkg_name, size_t* svc_id) {
   int len = strlen(pkg_name);
 
   allow_ro_return_t prev = allow_readonly(IPC_DRIVER_NUM, 0, pkg_name, len);
@@ -19,7 +19,7 @@ int ipc_discover(const char* pkg_name, int* svc_id) {
 
 int ipc_register_service_callback(const char *pkg_name,
                                   subscribe_upcall callback, void *ud) {
-  int svc_id;
+  size_t svc_id;
 
   // Look up the service id so we can subscribe as the service
   int ret = ipc_discover(pkg_name, &svc_id);
@@ -29,22 +29,22 @@ int ipc_register_service_callback(const char *pkg_name,
   return tock_subscribe_return_to_returncode(sval);
 }
 
-int ipc_register_client_callback(int svc_id, subscribe_upcall callback, void *ud) {
+int ipc_register_client_callback(size_t svc_id, subscribe_upcall callback, void *ud) {
   subscribe_return_t sval = subscribe(IPC_DRIVER_NUM, svc_id, callback, ud);
   return tock_subscribe_return_to_returncode(sval);
 }
 
-int ipc_notify_service(int pid) {
-  syscall_return_t res = command(IPC_DRIVER_NUM, 2, pid, 0);
+int ipc_notify_service(size_t pid) {
+  syscall_return_t res = command(IPC_DRIVER_NUM, 2, (int) pid, 0);
   return tock_command_return_novalue_to_returncode(res);
 }
 
-int ipc_notify_client(int pid) {
-  syscall_return_t res = command(IPC_DRIVER_NUM, 3, pid, 0);
+int ipc_notify_client(size_t pid) {
+  syscall_return_t res = command(IPC_DRIVER_NUM, 3, (int) pid, 0);
   return tock_command_return_novalue_to_returncode(res);
 }
 
-int ipc_share(int pid, void* base, int len) {
-  allow_rw_return_t aval = allow_readwrite(IPC_DRIVER_NUM, pid, base, len);
+int ipc_share(size_t pid, void* base, int len) {
+  allow_rw_return_t aval = allow_readwrite(IPC_DRIVER_NUM, (int) pid, base, len);
   return tock_allow_rw_return_to_returncode(aval);
 }
